ref(state): move state trees down a level
This is to make space for higher level state driver.
