+++ crates/pathfinder/src/state.rs
@@ -1,80 +1,3 @@
-//! Contains the [GlobalStateTree] and [ContractsStateTree] trees, which combined
-//! store the total StarkNet state.
-//!
-//! These are abstractions built-on the [Binary Merkle-Patricia Tree](MerkleTree).
-
-use rusqlite::Transaction;
-
-use crate::{
-    core::{
-        ContractAddress, ContractRoot, ContractStateHash, GlobalRoot, StorageAddress, StorageValue,
-    },
-    state::merkle_tree::MerkleTree,
-};
-
 mod merkle_node;
 mod merkle_tree;
-
-/// A Binary Merkle-Patricia Tree which contains
-/// the storage state of all StarkNet contracts.
-struct ContractsStateTree<'a> {
-    tree: MerkleTree<'a>,
-}
-
-impl<'a> ContractsStateTree<'a> {
-    pub fn load(transaction: &'a Transaction, root: ContractRoot) -> anyhow::Result<Self> {
-        // TODO: move the string into storage.
-        let tree = MerkleTree::load("tree_contracts".to_string(), transaction, root.0)?;
-
-        Ok(Self { tree })
-    }
-
-    pub fn get(&self, address: StorageAddress) -> anyhow::Result<StorageValue> {
-        let value = self.tree.get(address.0)?;
-        Ok(StorageValue(value))
-    }
-
-    pub fn set(&mut self, address: StorageAddress, value: StorageValue) -> anyhow::Result<()> {
-        self.tree.set(address.0, value.0)
-    }
-
-    /// Applies and persists any changes. Returns the new tree root.
-    pub fn apply(self) -> anyhow::Result<ContractRoot> {
-        let root = self.tree.commit()?;
-        Ok(ContractRoot(root))
-    }
-}
-
-/// A Binary Merkle-Patricia Tree which contains
-/// the global state of StarkNet.
-struct GlobalStateTree<'a> {
-    tree: MerkleTree<'a>,
-}
-
-impl<'a> GlobalStateTree<'a> {
-    pub fn load(transaction: &'a Transaction, root: GlobalRoot) -> anyhow::Result<Self> {
-        // TODO: move the string into storage.
-        let tree = MerkleTree::load("tree_global".to_string(), transaction, root.0)?;
-
-        Ok(Self { tree })
-    }
-
-    pub fn get(&self, address: ContractAddress) -> anyhow::Result<ContractStateHash> {
-        let value = self.tree.get(address.0)?;
-        Ok(ContractStateHash(value))
-    }
-
-    pub fn set(
-        &mut self,
-        address: ContractAddress,
-        value: ContractStateHash,
-    ) -> anyhow::Result<()> {
-        self.tree.set(address.0, value.0)
-    }
-
-    /// Applies and persists any changes. Returns the new global root.
-    pub fn apply(self) -> anyhow::Result<GlobalRoot> {
-        let root = self.tree.commit()?;
-        Ok(GlobalRoot(root))
-    }
-}
+mod state_tree;
